requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
base58==2.1.1
bitarray==2.7.3
cached-property==1.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
cytoolz==0.12.1
debugpy==1.6.6
decorator==5.1.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.3.3
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==2.3.0
eth-utils==1.10.0
executing==1.2.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
lru-dict==1.1.8
matplotlib-inline==0.1.6
multiaddr==0.0.9
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
netaddr==0.8.0
packaging==23.0
parsimonious==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-ecc==6.0.0
pycodestyle==2.10.0
pycryptodome==3.17
Pygments==2.14.0
pyrsistent==0.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
rlp==3.0.0
six==1.16.0
ssz==0.3.0
stack-data==0.6.2
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
urllib3==1.26.15
varint==1.0.2
wcwidth==0.2.6
web3==6.0.0
websockets==10.4
yarl==1.8.2
